Give expertise its due recognition.
JPEG stands for Joint Photographic Experts Group, but blahajfs only
previously checked for filenames ending in .jpg, which seems to imply
that the creators of the image standard may have been a Joint
Photographic Group of people who were not experts. But they deserve to
be honoured as the experts that they are so check for .jpeg too.

As a measure for backward compatibility with previous versions of
blahajfs and also for compatibility with a large amount of software
which has apparently deemed that photographic expertise is not to be
celebrated, also continue to accept the previous name pattern as well.